client: return existing members list on add_member
Returning just the newly added member provides no additional benefit to
the caller: the caller already knows which member it is adding, and is
left without any clue as to the remaining cluster.

In contrast, when the same command is run through 'etcdctl', one ends up
with a neat list of existing members that can then be provided to the
newly added member.

This patch aims at providing this additional context that is so much
needed in some situations.

Signed-off-by: Joao Eduardo Luis <joao@suse.com>